requirements.txt
streamlit==1.36.0
pdfplumber==0.11.4
python-docx==1.1.2
pytz==2024.1
